pyproject.toml
[tool.poetry]
name = "xcpcio-board-spider"
version = "0.0.16"
description = "XCPCIO Board Spider"
authors = ["Dup4 <lyuzhi.pan@gmail.com>"]
license = "MIT"
readme = "README.md"
classifiers = [
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
pyexecjs = "^1.5.1"
grequests = "^0.6.0"
gevent = "^22.10.2"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.2"
html5lib = "^1.1"
xlrd = "^2.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
flake8 = "^6.0.0"
pyright = "^1.1.305"
autopep8 = "^2.0.2"
pytest-snapshot = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
